u64 constructor tag (#65)
* Fix tag size

* Rename constructor index tag

* Update inkwell

+++ ssf-llvm/src/type_compiler.rs
@@ -63,7 +63,7 @@ impl<'c> TypeCompiler<'c> {
     pub fn compile_algebraic(
         &self,
         algebraic: &ssf::types::Algebraic,
-        index: Option<usize>,
+        index: Option<u64>,
     ) -> inkwell::types::StructType<'c> {
         let mut elements = vec![];
